refactor: extract LocateBinary into utility.h to remove duplication
Move the shared binary-locating logic from LocateDataGenerator and
LocateOptimizeGenerator into a common LocateBinary function in utility.h.
